fix: resolve all clippy warnings with --all-targets, improve mutex handling (#271)
* fix: resolve all clippy warnings with --all-targets, improve mutex handling

Fix 15+ clippy warnings that only appear with --all-targets (test code):
- field_reassign_with_default in auth store test
- type_complexity in lua test helper (allow)
- useless_vec in mcp client test
- bool_assert_comparison in storage test
- needless_borrow in provider api_types test
- clone_on_copy in circuit_breaker test
- unnecessary_cast in provider client test (2x)
- unnecessary_get_then_check in agent_loop test
- approx_constant in embedding test (3x, use FRAC_1_SQRT_2)
- len_zero in guardrails test
- impossible_comparisons in quality test
- manual_range_contains in tui tests (2x)
- map_or → is_some_and in chat_widget tests (4x)

Production code:
- Replace 6 mutex .unwrap() with .expect() in lua session API

* fix: return mlua::Error instead of panicking on mutex poison in lua API

Replace .expect() with .map_err(mlua::Error::external) to avoid
panicking across the FFI boundary in mlua callbacks. Addresses
greptile P2 review — mutex poison errors now propagate as Lua
errors instead of unwinding through C frames.

+++ crates/genesis-provider/src/client.rs
@@ -1253,9 +1253,7 @@ mod tests {
     #[test]
     fn backoff_delay_caps_at_max() {
         let d10 = backoff_delay(10);
-        assert!(
-            d10.as_millis() <= MAX_DELAY.as_millis() as u128 + MAX_DELAY.as_millis() as u128 / 4
-        );
+        assert!(d10.as_millis() <= MAX_DELAY.as_millis() + MAX_DELAY.as_millis() / 4);
     }
 
     #[test]
